LPS-112372 Rename com.liferay.info.item.fields.reader as com.liferay.info.item.field.reader since the classes inside represent a single field and not several
