+++ modules/apps/info/info-api/bnd.bnd
@@ -11,7 +11,7 @@ Export-Package:\
 	com.liferay.info.field,\
 	com.liferay.info.field.type,\
 	com.liferay.info.item,\
-	com.liferay.info.item.fields.reader,\
+	com.liferay.info.item.field.reader,\
 	com.liferay.info.item.provider,\
 	com.liferay.info.item.renderer,\
 	com.liferay.info.item.renderer.template,\
