+++ modules/apps/info/info-api/src/main/resources/com/liferay/info/item/field/reader/packageinfo
